+++ node/npm/glide/index.ts
@@ -74,8 +74,8 @@ function loadNativeBinding() {
 function initialize() {
     const nativeBinding = loadNativeBinding();
     const {
-        BitOffsetOptions,
         BitmapIndexType,
+        BitOffsetOptions,
         ConditionalChange,
         GeoAddOptions,
         GeospatialData,
@@ -126,8 +126,8 @@ function initialize() {
     } = nativeBinding;
 
     module.exports = {
-        BitOffsetOptions,
         BitmapIndexType,
+        BitOffsetOptions,
         ConditionalChange,
         GeoAddOptions,
         GeospatialData,