+++ node/src/commands/BitOffsetOptions.ts
@@ -5,19 +5,7 @@
 // Import below added to fix up the TSdoc link, but eslint blames for unused import.
 /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
 import { BaseClient } from "src/BaseClient";
-
-/**
- * Enumeration specifying if index arguments are BYTE indexes or BIT indexes.
- * Can be specified in {@link BitOffsetOptions}, which is an optional argument to the {@link BaseClient.bitcount|bitcount} command.
- *
- * since - Valkey version 7.0.0.
- */
-export enum BitmapIndexType {
-    /** Specifies that indexes provided to {@link BitOffsetOptions} are byte indexes. */
-    BYTE = "BYTE",
-    /** Specifies that indexes provided to {@link BitOffsetOptions} are bit indexes. */
-    BIT = "BIT",
-}
+import { BitmapIndexType } from "./BitmapIndexType";
 
 /**
  * Represents offsets specifying a string interval to analyze in the {@link BaseClient.bitcount|bitcount} command. The offsets are